shopinvader_product_attribute_set: fix random data conversion
When not using strict types the value coming from Odoo records will be forcedly converted
to the 1st type matching in the list of allowed types.

So, if the 1st matching type is str and the value can be converted to a string (eg: a float)
it will be converted to a string even if you expect to have a number in the final value.